WIP: TestChangesDefaultRules
