+++ src/logger/NetHttpClientLogger_test.go
@@ -12,7 +12,7 @@ func TestNetHttpClientLogger(t *testing.T) {
 
 	testLogger.SetLogFlag(false)
 
-	_, err := testLogger.Get(testHelper.DEMO_URL)
+	_, err := testLogger.Get(testHelper.demoURL)
 	if err == nil {
 		fmt.Println("Get request success")
 	}
