+++ src/logger/helper_test.go
@@ -4,10 +4,9 @@ import (
 	"testing"
 )
 
-
 func TestHelper(t *testing.T) {
 	testHelper := NewTestHelper()
-	if testHelper.DEMO_URL == "" {
+	if testHelper.demoURL == "" {
 		t.Error("Helper DEMO_URL is empty")
 	}
 }
